Refactors as messages/message-outbound config
