--- config/lib/middleware/messages/support/message-outbound.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = {
-  messageDirection: 'outbound-api-send',
-  shouldPostToPlatform: true,
-};
